Create Grading Calculator
The application will prompt the user to input the scale for which they are graded on. Next, they
will be prompted to enter their grades in one specific course which will be used to calculate their
average grade.
After calculating the average grade in that course, they will be asked if the course is weighted. If
so, they will be prompted to input the weighted amount. Then the application will calculate based
on weight.
Finally, the application will take all the current course grades of the user and calculate the GPA.
Once the average grades and GPA are calculated, they will be displayed on the screen